rename tile sizes (small*, medium, large, wide)
*we dont have any small ones yet.
