+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/BtcPrice/BtcPriceLargeTileView.axaml
@@ -5,7 +5,7 @@
              xmlns:vm="using:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
              mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="330"
              x:CompileBindings="True" x:DataType="vm:BtcPriceTileViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.BtcPrice.BtcPriceNormalTileView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.BtcPrice.BtcPriceLargeTileView">
   <DockPanel>
     <TextBlock Text="BTC Price" DockPanel.Dock="Top" />
     <StackPanel DockPanel.Dock="Left" Orientation="Horizontal" Spacing="5" VerticalAlignment="Center">
