+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/BtcPrice/BtcPriceMediumTileView.axaml
@@ -5,7 +5,7 @@
              xmlns:vm="using:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
              mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="150"
              x:CompileBindings="True" x:DataType="vm:BtcPriceTileViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.BtcPrice.BtcPriceSmallTileView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.BtcPrice.BtcPriceMediumTileView">
   <DockPanel>
     <TextBlock Text="BTC Price" DockPanel.Dock="Top" />
     <StackPanel DockPanel.Dock="Left" Orientation="Horizontal" Spacing="5" VerticalAlignment="Center">
