+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/RoundStatus/RoundStatusLargeTileView.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:vm="using:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
-             mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="150"
+             mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="330"
              x:CompileBindings="True" x:DataType="vm:RoundStatusTileViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.RoundStatus.RoundStatusSmallTileView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.RoundStatus.RoundStatusLargeTileView">
   <DockPanel>
     <TextBlock Text="Mixing Status" DockPanel.Dock="Top" />
     <Carousel Margin="0 10" VerticalAlignment="Center" SelectedIndex="{Binding SelectedIndex}">
