+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/RoundStatus/RoundStatusMediumTileView.axaml
@@ -3,9 +3,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:vm="using:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
-             mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="330"
+             mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="150"
              x:CompileBindings="True" x:DataType="vm:RoundStatusTileViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.RoundStatus.RoundStatusNormalTileView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.RoundStatus.RoundStatusMediumTileView">
   <DockPanel>
     <TextBlock Text="Mixing Status" DockPanel.Dock="Top" />
     <Carousel Margin="0 10" VerticalAlignment="Center" SelectedIndex="{Binding SelectedIndex}">
