+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/WalletBalance/WalletBalanceLargeTileView.axaml
@@ -5,7 +5,7 @@
              xmlns:vm="using:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
              mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="330"
              x:CompileBindings="True" x:DataType="vm:WalletBalanceTileViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletBalance.WalletBalanceNormalTileView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletBalance.WalletBalanceLargeTileView">
   <DockPanel>
     <TextBlock Text="Balance" DockPanel.Dock="Top" />
     <StackPanel DockPanel.Dock="Left" Orientation="Horizontal" Spacing="16">
