+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/WalletBalanceChartTileView.axaml
@@ -9,9 +9,9 @@
              x:CompileBindings="True" x:DataType="vm:WalletBalanceChartTileViewModel"
              x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletBalanceChartTileView">
   <controls:TileControl TileSize="{Binding CurrentTileSize}">
-    <wbc:WalletBalanceChartSmallTileView />
+    <wbc:WalletBalanceChartMediumTileView />
     <controls:TileControl.LargeSizeContent>
-      <wbc:WalletBalanceChartNormalTileView />
+      <wbc:WalletBalanceChartLargeTileView />
     </controls:TileControl.LargeSizeContent>
     <controls:TileControl.WideSizeContent>
       <wbc:WalletBalanceChartWideTileView />
