+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/WalletPieChart/WalletPieChartLargeTileView.axaml
@@ -6,13 +6,13 @@
              xmlns:controls="clr-namespace:WalletWasabi.Fluent.Controls"
              xmlns:wallets="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles"
              xmlns:converters="clr-namespace:WalletWasabi.Fluent.Converters"
-             mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="150"
+             mc:Ignorable="d" d:DesignWidth="330" d:DesignHeight="330"
              x:CompileBindings="True" x:DataType="vm:WalletPieChartTileViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletPieChart.WalletPieChartSmallTileView">
+             x:Class="WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletPieChart.WalletPieChartLargeTileView">
   <DockPanel>
     <TextBlock Text="Wallet Privacy Status" DockPanel.Dock="Top" />
     <DockPanel LastChildFill="True" TextBlock.FontSize="12"
-               VerticalAlignment="Center">
+               VerticalAlignment="Stretch">
       <ItemsRepeater DockPanel.Dock="Bottom" Margin="0,0,0,-8" Items="{Binding TestDataPointsLegend}">
         <ItemsRepeater.ItemTemplate>
           <DataTemplate x:DataType="wallets:DataLegend">
@@ -32,6 +32,9 @@
           <StackLayout />
         </ItemsRepeater.Layout>
       </ItemsRepeater>
+      <Panel Margin="8" VerticalAlignment="Center" HorizontalAlignment="Center">
+        <controls:RingChartControl DataPoints="{Binding TestDataPoints}" />
+      </Panel>
     </DockPanel>
   </DockPanel>
 </UserControl>
