+++ WalletWasabi.Fluent/Views/Wallets/Home/Tiles/WalletPieChart/WalletPieChartLargeTileView.axaml.cs
@@ -2,11 +2,11 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletBalance
+namespace WalletWasabi.Fluent.Views.Wallets.Home.Tiles.WalletPieChart
 {
-	public class WalletBalanceNormalTileView : UserControl
+	public class WalletPieChartLargeTileView : UserControl
 	{
-		public WalletBalanceNormalTileView()
+		public WalletPieChartLargeTileView()
 		{
 			InitializeComponent();
 		}
